Se modularizo, creando un common llamado InputsForm que recibe por props los datos necesarios para el input.
